requirements.txt
flask~=2.3.2
pytesseract~=0.3.10
pyscreenshot~=3.1
Pillow~=9.5.0
yolov5~=7.0.12